+++ src/pv/transfer.c
@@ -402,12 +402,21 @@ static int pv__transfer_read(pvstate_t state, int fd, int *eof_in, int *eof_out,
 		return 1;
 	}
 
-	if (state->read_errors_in_a_row < 10) {
-		amount_to_skip = state->read_errors_in_a_row < 5 ? 1 : 2;
-	} else if (state->read_errors_in_a_row < 20) {
-		amount_to_skip = 1 << (state->read_errors_in_a_row - 10);
-	} else {
-		amount_to_skip = 512;
+	/*
+	 * If a non-zero error skip block size was given, just use that,
+	 * otherwise start small and ramp up based on the number of errors
+	 * in a row.
+	 */
+	if (state->error_skip_block > 0) {
+		amount_to_skip = state->error_skip_block;
+		} else {
+		if (state->read_errors_in_a_row < 10) {
+			amount_to_skip = state->read_errors_in_a_row < 5 ? 1 : 2;
+		} else if (state->read_errors_in_a_row < 20) {
+			amount_to_skip = 1 << (state->read_errors_in_a_row - 10);
+		} else {
+			amount_to_skip = 512;
+		}
 	}
 
 	/*